backEnd/requirements.txt
fastapi
Jinja2
loguru
passlib
PyMySQL
python-multipart
redis
requests
SQLAlchemy
uvicorn
python-jose[cryptography]
aliyun-log-python-sdk
httpx
aiofiles
chardet
gitpython